requirements.txt
django-mptt
django-imagekit
git+https://github.com/AliYmn/django-ckeditor